Fix: Dto 타입 변경

+++ src/main/java/noodlezip/store/dto/MenuRequestDto.java
@@ -17,7 +17,7 @@ public class MenuRequestDto {
     private String menuImageUrl;
 
     private Long ramenCategoryId; // ramenCategory 엔티티의 ID
-    private Integer ramenSoupId;     // ramenSoup 엔티티의 ID
+    private Long ramenSoupId;     // ramenSoup 엔티티의 ID
 
     private List<Long> defaultToppingIds; // 기본 토핑 ID 목록 (다대다 관계)
 }
